requirements.txt
streamlit
pillow
numpy
tensorflow-macos==2.17.0
tensorflow-metal==1.1.0

